requirements.txt
fastapi==0.61.1
feedparser==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
uvicorn==0.12.1
